Audit: Consolidate ModActivation.ts into ModSquashBehavioural.ts (#1769)
Move the constant-neuron test from the separate ModActivation.ts file into
ModSquashBehavioural.ts where the other ModSquash/ModActivation behavioural
tests live. Improves test name from "ModActivation-Constant" to
"ModSquash: modifies non-constant neurons when constant neurons are present".
